app/qr_code.py
import qrcode
import io
from flask import url_for, send_file
from PIL import Image 

def generate_qr_code(data: str, image_factory=None, **kwargs) -> bytes:
    """Генерирует QR-код и возвращает его в виде байтовой строки."""

    qr = qrcode.QRCode(
        version=None,
        error_correction=qrcode.constants.ERROR_CORRECT_H,
        box_size=10,
        border=4,
        image_factory=image_factory,
    )
    qr.add_data(data)
    qr.make(fit=True)

    img = qr.make_image(fill_color="black", back_color="white").convert('RGBA')

    # Убираем фон
    datas = img.getdata()
    newData = []
    for item in datas:
        if item[0] == 255 and item[1] == 255 and item[2] == 255:
            newData.append((255, 255, 255, 0))
        else:
            newData.append(item)

    img.putdata(newData)

    # Сохраняем QR-код в буфер в памяти
    buffer = io.BytesIO()
    img.save(buffer, format="PNG")
    buffer.seek(0)

    return buffer.getvalue()