requirements.txt
Flask
Flask-SQLAlchemy
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
oauth2client
psycopg2-binary
python-dotenv
requests
SQLAlchemy
Flask[async]
flask-migrate
flask-adminlte3
qrcode
Pillow
flask-login
flask-wtf
WTForms-SQLAlchemy
gunicorn